refactor(sync): extract pure helpers from CloudSyncService
- CloudSyncEncryption.kt: encryptData/decryptData/getOrCreateSecretKey/getCipher
- CloudSyncSerialization.kt: serializeToJson/deserializeFromJson
- CloudSyncMergeHelpers.kt: identifyConflicts/mergeReadingPosition/mergeAnnotation

These methods don't reference class instance state and lift cleanly into
top-level functions in the same package. The conflict-resolution
methods (resolveConflictsAutomatically/Local/Remote/merge) remain in
the class because they touch the StateFlow + applyRemoteChange method.

CloudSyncService.kt: 884 -> 733 LOC.
